docker-requirements.txt
bosdyn-api>=5.0.0
bosdyn-client>=5.0.0
bosdyn-core>=5.0.0
certifi>=2024.2.2
charset-normalizer>=3.3.2
deprecated>=1.2.14
grpcio>=1.62.0
idna>=3.6
numpy>=1.26.4
protobuf>=4.25.3
pyjwt>=2.8.0
pynmea2>=1.20.0
requests>=2.31.0
urllib3>=2.2.1
wrapt>=1.16.0